refactor(api/v2): :art: Move application logic to service
Moves the application logic to the application service to have better separation of concerns.
